add interface for qp solver selection
