+++ SoftRobots.InverseConfig.cmake.in
@@ -10,6 +10,8 @@ find_package(Sofa.Component.SolidMechanics.FEM.Elastic QUIET REQUIRED)
 find_package(libqpOASES REQUIRED)
 include_directories(${OASES_INCLUDE_DIRS})
 
+find_package(proxsuite REQUIRED)
+
 if(NOT TARGET @PROJECT_NAME@)
     include("${CMAKE_CURRENT_LIST_DIR}/@PROJECT_NAME@Targets.cmake")
 endif()
