+++ src/SoftRobots.Inverse/component/solver/QPInverseProblemSolver.cpp
@@ -52,6 +52,8 @@
 #include <SoftRobots.Inverse/component/solver/QPInverseProblemSolver.h>
 #include <SoftRobots.Inverse/component/solver/modules/QPMechanicalSetConstraint.h>
 #include <SoftRobots.Inverse/component/solver/modules/QPMechanicalAccumulateConstraint.h>
+#include <SoftRobots.Inverse/component/solver/modules/QPInverseProblemQPOases.h>
+#include <SoftRobots.Inverse/component/solver/modules/QPInverseProblemProxQP.h> // TODO
 
 using sofa::simulation::mechanicalvisitor::MechanicalProjectJacobianMatrixVisitor;
 using sofa::simulation::mechanicalvisitor::MechanicalResetConstraintVisitor;
@@ -122,6 +124,8 @@ QPInverseProblemSolver::QPInverseProblemSolver()
 
     , d_responseFriction(initData(&d_responseFriction, 0., "responseFriction", "Response friction for contact resolution"))
 
+    , d_qpSolver(initData(&d_qpSolver, "qpSolver", "QP solver implementation to be used"))
+
     , d_epsilon(initData(&d_epsilon, 1e-3, "epsilon",
                          "An energy term is added in the minimization process. \n"
                          "Epsilon has to be chosen sufficiently small so that the deformation \n"
@@ -152,13 +156,36 @@ QPInverseProblemSolver::QPInverseProblemSolver()
 {
     createProblems();
     d_graph.setWidget("graph");
+
+    sofa::helper::OptionsGroup qpSolvers{"qpOASES" , "proxQP"};
+    qpSolvers.setSelectedItem(QPSolverImpl::QPOASES);
+    d_qpSolver.setValue(qpSolvers);
 }
 
 void QPInverseProblemSolver::createProblems()
 {
-    m_CP1 = new module::QPInverseProblemImpl();
-    m_CP2 = new module::QPInverseProblemImpl();
-    m_CP3 = new module::QPInverseProblemImpl();
+    switch(d_qpSolver.getValue().getSelectedId())
+    {
+    case QPSolverImpl::QPOASES :
+        msg_info() << "Using qpOASES solver";
+        m_CP1 = new module::QPInverseProblemQPOases();
+        m_CP2 = new module::QPInverseProblemQPOases();
+        m_CP3 = new module::QPInverseProblemQPOases();
+        break;
+
+    case QPSolverImpl::PROXQP :
+        // TODO conditionnal compilation w.r.t. optionnal proxQP dependency
+        msg_info() << "Using proxQP solver";
+        m_CP1 = new module::QPInverseProblemProxQP();
+        m_CP2 = new module::QPInverseProblemProxQP();
+        m_CP3 = new module::QPInverseProblemProxQP();
+        break;
+
+    default :
+        // TODO
+        break;
+    }
+
 
     m_currentCP = m_CP1;
 }
