+++ src/SoftRobots.Inverse/component/solver/modules/QPInverseProblemProxQP.h
@@ -29,20 +29,8 @@
 #pragma once
 
 #include <Eigen/Core>
-#include <qpOASES/Types.hpp>
 
-//When the macro is defined by qpOASES/Types.hpp it has side effects on the
-//rest of the code, in particular in boost (source_location). This macro is
-//no longer necessary if a modern compiler is used.
-#ifdef snprintf
-    #undef snprintf
-#endif
-
-#include <qpOASES/QProblem.hpp>
-#include <SoftRobots/component/behavior/SoftRobotsBaseConstraint.h>
-
-#include <SoftRobots.Inverse/component/solver/modules/ConstraintHandler.h>
-#include <SoftRobots.Inverse/component/solver/modules/QPInverseProblem.h>
+#include <SoftRobots.Inverse/component/solver/modules/QPInverseProblemImpl.h>
 
 #include <SoftRobots.Inverse/component/config.h>
 
@@ -51,59 +39,23 @@ namespace softrobotsinverse::solver::module
 {
 
 using sofa::type::vector;
-using qpOASES::real_t;
 
-
-class SOFA_SOFTROBOTS_INVERSE_API QPInverseProblemImpl : public QPInverseProblem
+class SOFA_SOFTROBOTS_INVERSE_API QPInverseProblemProxQP : public QPInverseProblemImpl
 {
 
 public:
-
-    typedef Eigen::Ref<Eigen::MatrixXd> RefMat;
-    typedef const Eigen::Ref<const Eigen::MatrixXd> ConstRefMat;
-    typedef Eigen::Ref<Eigen::VectorXd> RefVec;
-    typedef const Eigen::Ref<const Eigen::VectorXd> ConstRefVec;
-
-    QPInverseProblemImpl();
-    virtual ~QPInverseProblemImpl();
-
-    void init();
-    void solve(double &objective, int &iterations);
-    void setMinContactForces(const double& minContactForces) {m_qpCParams->minContactForces = minContactForces; m_qpCParams->hasMinContactForces = true;}
-    void setMaxContactForces(const double& maxContactForces) {m_qpCParams->maxContactForces = maxContactForces; m_qpCParams->hasMaxContactForces = true;}
+    QPInverseProblemProxQP() = default;
+    virtual ~QPInverseProblemProxQP() = default;
 
 protected:
-
-    ConstraintHandler* m_constraintHandler;
-    ConstraintHandler::QPConstraintParams* m_qpCParams;
-
-    // Utils to prevent cycling in pivot algorithm
-    vector<int>   m_currentSequence;
-    vector<int>   m_previousSequence;
-    vector<int>   m_sequence;
-
-    int m_iteration{0};
-    int m_step{0};
-
-    // must be reimplemented in each qp solver specific child class
     virtual void solveInverseProblem(double &objective,
                              vector<double> &result,
-                             vector<double> &dual) = 0;
+                             vector<double> &dual) override;
 
-    void computeEnergyWeight(double& weight);
-    void buildQPMatrices();
-
-    void solveWithContact(vector<double>& result, double &objective, int &iterations);
-    void solveContacts(vector<double>& res);
-
-    void updateLambda(const vector<double>& x);
-    bool isFeasible(const vector<double>& x);
-
-    bool checkAndUpdatePivot(const vector<double>&result, const vector<double>&dual);
-    bool isCycling(const int pivot);
-    double getDelta(const vector<double> &result, const int& index);
-    bool isIn(const vector<int> list, const int elem);
-    std::string getContactsState();
+private:
+    void updateQPMatrices(Eigen::MatrixXd& H, Eigen::VectorXd& g, Eigen::MatrixXd& A,
+       Eigen::VectorXd& b, Eigen::MatrixXd& C, Eigen::VectorXd& l, Eigen::VectorXd& u, 
+       Eigen::VectorXd& lbox, Eigen::VectorXd& ubox) const;
 };
 
 } //namespace
